requirements.txt
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.4
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
docopt==0.6.2
EasyProcess==1.1
entrypoint2==1.1
executing==1.2.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.12.2
jedi==0.19.0
jeepney==0.8.0
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mss==9.0.1
nest-asyncio==1.5.8
nltk==3.6.1
numpy==1.24.4
packaging==23.1
pandas==1.5.2
pandas-stubs==1.2.0.10
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.11
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
pyzmq==25.1.1
regex==2023.8.8
requests==2.28.1
six==1.16.0
sounddevice==0.4.6
srt==3.5.3
stack-data==0.6.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.8.0
urllib3==1.26.16
vosk==0.3.45
wcwidth==0.2.6
websockets==11.0.3
yarg==0.1.9
zipp==3.17.0
